Separate Phantom Deep Linking code from view model
